billing_module/billing/requirements.txt
uvicorn==0.24.0
fastapi==0.100.1
gunicorn==21.2.0
python-dateutil==2.9.0
yookassa==3.1.0
pydantic-settings==2.2.1
sqlalchemy==2.0.29
asyncpg==0.29.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
python-logstash==0.4.8
python-logstash-async==3.0.0
pyjwt==2.8.0
aiohttp==3.9.1